+++ src/main.rs
@@ -9,7 +9,7 @@ extern crate libxch;
 use clap::{Arg, App};
 const VERSION: &'static str = env!("CARGO_PKG_VERSION");
 const AUTHORS: &'static str = env!("CARGO_PKG_AUTHORS");
-use libxch::xch;
+use libxch::{xch, xch_non_atomic};
 
 fn main() {
     let matches = App::new("xch")
@@ -22,12 +22,22 @@ fn main() {
         .arg(Arg::with_name("PATH2")
             .help("The other path to exchange")
             .required(true))
+        .arg(Arg::with_name("non-atomic")
+            .long("non-atomic")
+            .short("n")
+            .help("Use non atomic exchange if atomic is not available"))
         .get_matches();
 
     let path1 = matches.value_of("PATH1").expect("clap should have covered this");
     let path2 = matches.value_of("PATH2").expect("clap should have covered this");
+    let non_atomic = matches.is_present("non-atomic");
 
-    let exit_code = match xch(path1, path2) {
+    let xch_result = if non_atomic {
+        xch_non_atomic(path1, path2)
+    } else {
+        xch(path1, path2)
+    };
+    let exit_code = match xch_result {
         Ok(_) => 0,
         Err(e) => {
             eprintln!("Could not swap files: {}", e);